refactor: switch to single playbook composed of many partials
This is the first step of making Ansible work for us as a single
playbook configured via variables.

Also, this commit:

1. adds preamble and epilogue for marking start and finish of play.
2. revisits documentation.
3. remove cloud-init related scripts and templates.

There is one issue with the telegraf for the initial run. We will need
to address it.

As a side note, I am thinking more of a "teloshost" role instead of a
playbook. Let's see how it rolls.